Added update methods

+++ src/main/java/com/Snowhite/domain/Inventory.java
@@ -15,7 +15,7 @@ public class Inventory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(nullable = false)
     private long inventoryNumber;
@@ -61,5 +61,5 @@ public class Inventory {
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private Status status;
+    private Status status ;
 }
